refactor: move import profiles to router view enhancement: initial app selection view
