+++ packages/desktop/views/login/import-third-party-profiles/import-third-party-profiles-route.enum.ts
@@ -1,0 +1,4 @@
+export enum ImportThirdPartyProfilesRoute {
+    SelectApps = 'selectApps',
+    ImportProfiles = 'importProfiles',
+}
