+++ packages/desktop/views/login/import-third-party-profiles/import-third-party-profiles.router.ts
@@ -1,0 +1,23 @@
+import { appRouter, Subrouter } from '@core/router'
+import { get, writable } from 'svelte/store'
+import { ImportThirdPartyProfilesRoute } from './import-third-party-profiles-route.enum'
+
+export const importThirdPartyProfilesRoute = writable<ImportThirdPartyProfilesRoute>(undefined)
+export const importThirdPartyProfilesRouter = writable<ImportThirdPartyProfilesRouter>(undefined)
+
+export class ImportThirdPartyProfilesRouter extends Subrouter<ImportThirdPartyProfilesRoute> {
+    constructor() {
+        super(ImportThirdPartyProfilesRoute.SelectApps, importThirdPartyProfilesRoute, get(appRouter))
+    }
+
+    next(): void {
+        const currentRoute = get(this.routeStore)
+        switch (currentRoute) {
+            case ImportThirdPartyProfilesRoute.SelectApps:
+                this.setNext(ImportThirdPartyProfilesRoute.ImportProfiles)
+                break
+            case ImportThirdPartyProfilesRoute.ImportProfiles:
+                break
+        }
+    }
+}
