+++ packages/desktop/views/login/import-third-party-profiles/index.ts
@@ -1,0 +1,4 @@
+export { default as SendFlowRouterView } from './ImportThirdPartyProfilesRouterView.svelte'
+
+export * from './import-third-party-profiles-route.enum'
+export * from './import-third-party-profiles.router'
