+++ packages/desktop/views/login/import-third-party-profiles/views/index.ts
@@ -1,0 +1,2 @@
+export { default as ImportProfilesView } from './ImportProfilesView.svelte'
+export { default as SelectAppsView } from './SelectAppsView.svelte'
